feat: Add integrations for Gemini, Ollama, and OpenAI APIs
- Implemented Gemini API integration with request and response handling.
- Added Ollama API integration for prompt completion.
- Created OpenAI API integration with support for chat completions and model listing.
- Introduced result types for engine configuration and response handling.
- Added configuration files for Kubex AI squad with detailed settings and guidelines.
- Included instructions for Kubex AI squad with project-specific details and governance.
