+++ internal/core/engine/history_manager.go
