+++ internal/types/result.go
